+++ ExpandWorldPrefabs/ObjectsFiltering.cs
@@ -13,11 +13,21 @@ public class ObjectsFiltering
   {
     if (objects.Length == 0) return [];
     var maxRadius = objects.Max(o => o.MaxDistance);
+    if (maxRadius > 10000)
+    {
+      var zdos = ZDOMan.instance.m_objectsByID.Values;
+      return GetObjects(limit, zdos, objects, zdo, parameters);
+    }
     var zdoLists = GetSectorIndices(zdo.m_position, maxRadius);
     return GetObjects(limit, zdoLists, objects, zdo, parameters);
   }
   public static ZDO[] GetNearby(int limit, Object objects, ZDO zdo, Dictionary<string, string> parameters)
   {
+    if (objects.MaxDistance > 10000)
+    {
+      var zdos = ZDOMan.instance.m_objectsByID.Values;
+      return GetObjects(limit, zdos, objects, zdo, parameters);
+    }
     var zdoLists = GetSectorIndices(zdo.m_position, objects.MaxDistance);
     return GetObjects(limit, zdoLists, objects, zdo, parameters);
   }
@@ -26,6 +36,15 @@ public class ObjectsFiltering
     var pos = zdo.m_position;
     var zm = ZDOMan.instance;
     var query = zdoLists.SelectMany(z => z).Where(z => objects.IsValid(z, pos, parameters));
+    if (limit > 0)
+      query = query.OrderBy(z => Utils.DistanceXZ(z.m_position, pos)).Take(limit);
+    return query.ToArray();
+  }
+  private static ZDO[] GetObjects(int limit, Dictionary<ZDOID, ZDO>.ValueCollection zdos, Object objects, ZDO zdo, Dictionary<string, string> parameters)
+  {
+    var pos = zdo.m_position;
+    var zm = ZDOMan.instance;
+    var query = zdos.Where(z => objects.IsValid(z, pos, parameters));
     if (limit > 0)
       query = query.OrderBy(z => Utils.DistanceXZ(z.m_position, pos)).Take(limit);
     return query.ToArray();
@@ -39,11 +58,30 @@ public class ObjectsFiltering
       query = query.OrderBy(z => Utils.DistanceXZ(z.m_position, pos)).Take(limit);
     return query.ToArray();
   }
+  private static ZDO[] GetObjects(int limit, Dictionary<ZDOID, ZDO>.ValueCollection zdos, Object[] objects, ZDO zdo, Dictionary<string, string> parameters)
+  {
+    var pos = zdo.m_position;
+    var zm = ZDOMan.instance;
+    var query = zdos.Where(z => objects.Any(o => o.IsValid(z, pos, parameters)));
+    if (limit > 0)
+      query = query.OrderBy(z => Utils.DistanceXZ(z.m_position, pos)).Take(limit);
+    return query.ToArray();
+  }
+
+
 
   public static bool HasNearby(Range<int>? limit, Object[] objects, ZDO zdo, Dictionary<string, string> parameters)
   {
     if (objects.Length == 0) return true;
     var maxRadius = objects.Max(o => o.MaxDistance);
+    if (maxRadius > 10000)
+    {
+      var zdos = ZDOMan.instance.m_objectsByID.Values;
+      if (limit == null)
+        return HasAllObjects(zdos, objects, zdo, parameters);
+      else
+        return HasLimitObjects(zdos, limit, objects, zdo, parameters);
+    }
     var zdoLists = GetSectorIndices(zdo.m_position, maxRadius);
     if (limit == null)
       return HasAllObjects(zdoLists, objects, zdo, parameters);
@@ -67,6 +105,12 @@ public class ObjectsFiltering
     var zm = ZDOMan.instance;
     return objects.All(o => zdoLists.Any(z => z.Any(z => o.IsValid(z, pos, parameters) && z != zdo)));
   }
+  private static bool HasAllObjects(Dictionary<ZDOID, ZDO>.ValueCollection zdos, Object[] objects, ZDO zdo, Dictionary<string, string> parameters)
+  {
+    var pos = zdo.m_position;
+    var zm = ZDOMan.instance;
+    return objects.All(o => zdos.Any(z => o.IsValid(z, pos, parameters) && z != zdo));
+  }
   private static bool HasLimitObjects(List<List<ZDO>> zdoLists, Range<int> limit, Object[] objects, ZDO zdo, Dictionary<string, string> parameters)
   {
     var pos = zdo.m_position;
@@ -85,6 +129,22 @@ public class ObjectsFiltering
     }
     return limit.Min <= counter && counter <= limit.Max;
   }
+
+  private static bool HasLimitObjects(Dictionary<ZDOID, ZDO>.ValueCollection zdos, Range<int> limit, Object[] objects, ZDO zdo, Dictionary<string, string> parameters)
+  {
+    var pos = zdo.m_position;
+    var counter = 0;
+    var useMax = limit.Max > 0;
+    foreach (var z in zdos)
+    {
+      var valid = objects.FirstOrDefault(o => o.IsValid(z, pos, parameters) && z != zdo);
+      if (valid == null) continue;
+      counter += valid.Weight;
+      if (useMax && limit.Max < counter) return false;
+      if (limit.Min <= counter && !useMax) return true;
+    }
+    return limit.Min <= counter && counter <= limit.Max;
+  }
   private static List<List<ZDO>> GetSectorIndices(Vector3 pos, float radius)
   {
     List<List<ZDO>> zdoLists = [];
@@ -100,13 +160,14 @@ public class ObjectsFiltering
         var index = zm.SectorToIndex(zone);
         if (index < 0 || index >= zm.m_objectsBySector.Length)
         {
-          if (zm.m_objectsByOutsideSector.TryGetValue(zone, out var list))
+          if (zm.m_objectsByOutsideSector.TryGetValue(zone, out var list) && list != null)
             zdoLists.Add(list);
           continue;
         }
         if (indices.Contains(index)) continue;
-        zdoLists.Add(zm.m_objectsBySector[index]);
         indices.Add(index);
+        if (zm.m_objectsBySector[index] != null)
+          zdoLists.Add(zm.m_objectsBySector[index]);
       }
     }
     return zdoLists;